# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a HUB API.",
        "title": "HUB API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/stat": {
            "get": {
                "description": "Get daily statistics for accounts, needles, and volumes using GET method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get statistics by GET",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of days to return (default: 30)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of daily statistics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Stat"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid count parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Get daily statistics for accounts, needles, and volumes using POST method",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Get statistics by POST",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of days to return (default: 30)",
                        "name": "count",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array of daily statistics",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Stat"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid count parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "types.Stat": {
            "type": "object",
            "properties": {
                "dailyAccounts": {
                    "description": "new created accounts at this day",
                    "type": "integer"
                },
                "dailyBuckets": {
                    "description": "new created buckets at this day",
                    "type": "integer"
                },
                "dailyNeedles": {
                    "description": "new created needles at this day",
                    "type": "integer"
                },
                "dailyVolumes": {
                    "description": "new created volumes at this day",
                    "type": "integer"
                },
                "day": {
                    "type": "string"
                },
                "totalAccounts": {
                    "description": "total accounts until this day",
                    "type": "integer"
                },
                "totalBuckets": {
                    "description": "total buckets until this day",
                    "type": "integer"
                },
                "totalNeedles": {
                    "description": "total needles until this day",
                    "type": "integer"
                },
                "totalVolumes": {
                    "description": "total volumes until this day",
                    "type": "integer"
                }
            }
        }
    }
}